chore(calculator): rename OrderItemButton -> ~Group
